feat: [Refactor P1-4] Extract shared CFG/RESPONSE_CFG builders used by the AWS and Cloudflare compilers (#185)
