+++ .devloop/ledger.jsonl
@@ -1,0 +1,1 @@
+{"version":1,"eventId":"evt_612b9917-93c6-467a-af46-eb80137fe52b","eventType":"takt_run_imported","timestamp":"2026-06-28T15:32:03.750Z","repoPath":"/Volumes/Satechi/Developer/cdn-security-framework","issueNumber":189,"runSlug":"20260628-150155-issue-189-refactor-p2-2-replac","taktRunPath":".takt/runs/20260628-150155-issue-189-refactor-p2-2-replac","status":"completed","task":"## Issue #189: [Refactor P2-2] Replace catch (e: any) with unknown + shared error helpers (64 sites)\n\n# [Refactor P2-2] catch (e: any) を unknown + 共有エラーヘルパーに置換(64箇所)\n\nPhase 2(TypeScript活用)/ 親issue: #176\n\n## 概要\n\n`tsconfig.json` は `useUnknownInCatchVariables: true` を設定しているのに、コード側が `catch (e: any)` と**明示annotationで安全機構を打ち消している**(src全体で53箇所+cli.tsの11箇所)。`unknown` 化し、`e.code` / `e.message` / `e.validationErrors` へのアクセスを小さな型ガードヘルパー経由にする。\n\n## 現状の根拠\n\n```bash\ngrep -rn \"catch (e: any)\" src --include=\"*.ts\" | wc -l   # → 53\n# cli.ts内のcatch-log-exit型: L3210, 3608, 3629, 3645, 3661, 3681, 3696, 3726, 3755, 3838, 3870\n```\n\n典型パターン3種:\n\n```ts\ncatch (e: any) { if (e.code === 'ENOENT') {...} console.error('...', e.message); process.exit(1); }  // ファイルIO\ncatch (e: any) { if (Array.isArray(e.validationErrors)) {...} }                                       // 自作検証エラー\ncatch (e: any) { console.error('[ERROR]', e.message); process.exit(1); }                              // CLI汎用\n```\n\n## 変更内容(設計)\n\n新規 `src/scripts/lib/errors.ts`:\n\n```ts\n/** unknown から安全にメッセージを取り出す(Errorでなければ String() に落とす) */\nexport function errorMessage(e: unknown): string {\n  return e instanceof Error ? e.message : String(e);\n}\n\n/** Node の ErrnoException 判定(e.code === 'ENOENT' 等の前段) */\nexport function isErrnoException(e: unknown): e is NodeJS.ErrnoException {\n  return e instanceof Error && typeof (e as NodeJS.ErrnoException).code === 'string';\n}\n\n/** validateAuthGates / validateOriginAuth が投げる validationErrors 付きエラー。\n *  現状は `const error: any = new Error(...); error.validationErrors = errors;`\n *  ([compile-core.ts:359-361](src/scripts/lib/compile-core.ts))という動的付与なので、専用クラスに置き換える。 */\nexport class PolicyValidationError extends Error {\n  constructor(message: string, public readonly validationErrors: string[]) { super(message); this.name = 'PolicyValidationError'; }\n}\nexport function isPolicyValidationError(e: unknown): e is PolicyValidationError;\n```\n\n置換例:\n\n```diff\n-  } catch (e: any) {\n-    if (e.code === 'ENOENT') {\n+  } catch (e: unknown) {\n+    if (isErrnoException(e) && e.code === 'ENOENT') {\n       console.error('Error: policy file not found:', policyPath);\n       process.exit(1);\n     }\n-    console.error('Error: failed to parse policy YAML:', e.message);\n+    console.error('Error: failed to parse policy YAML:', errorMessage(e));\n     process.exit(1);\n   }\n```\n\n### 注意点\n\n1. **`PolicyValidationError` への置換は投げる側と受ける側を同一コミットで**: 投げる側 [compile-core.ts:359-361, 494-496](src/scripts/lib/compile-core.ts)、受ける側 [policy-lint.ts:104-112](src/scripts/policy-lint.ts) と `lib/` 配下(`grep -rn \"validationErrors\" src` で全列挙してから着手)\n2. `errorMessage(e)` は `e.message` と完全同値ではない(非Errorが投げられた場合 `undefined` → `String(e)` になる)。実害のない改善だが、互換テストで例外系メッセージが変わらないことを確認する\n3. テストファイル内の `catch (e: any)`(`e.stack` 参照など)も対象。`e instanceof Error ? e.stack : String(e)` 形へ\n\n## 実装手順(TDD)\n\n1. `src/scripts/errors-unit-tests.ts` を新規作成し、ヘルパー4つの仕様をテスト(Error/非Error/code付き/validationErrors付き)。`test:unit` チェーンに追加\n2. `PolicyValidationError` 導入(投げる側+受ける側を同時に)→ `npm run lint:policy:ci` で**lint失敗時の出力文言が不変**なことを確認(わざと壊したポリシーで before/after 比較)\n3. 残りの `catch (e: any)` を領域ごとに置換(scripts/lib → scripts → lib/parser/validator/emitter → cli.ts → tests)。1領域=1コミット\n4. 各コミットで `npm run typecheck && npm run test:ci`\n\n## 受け入れ基準\n\n- [ ] `grep -rn \"catch (e: any)\" src --include=\"*.ts\"` が0件\n- [ ] `grep -rn \"error: any = new Error\\|err: any = new Error\" src` が0件(PolicyValidationErrorに置換済み)\n- [ ] エラー時のCLI出力文言が変わっていない(壊れたポリシーでの before/after をPRに貼付)\n- [ ] `npm run test:ci` 成功 / `npm run test:drift` 無変化\n\n## スコープ外\n\n- `: any` 全般の削減(P2-3)\n- エラーメッセージの改善・国際化\n\n## 依存関係\n\n- **P2-1(import化)の後**(unknown化で顕在化する型エラーがimport化済みだと正しく出るため)\n\n想定規模: **M**\n\n\n### Labels\npriority/p2, refactor, agent:ready","workflow":"default-subscription-devloop","startTime":"2026-06-28T15:01:55.099Z","endTime":"2026-06-28T15:31:58.394Z","artifacts":[{"kind":"log","path":".takt/runs/20260628-150155-issue-189-refactor-p2-2-replac/logs/20260629-000155-fpwbuq.jsonl","sha256":"f0d8c810aab1caee327ee922cb8aa77d85648ca30c7db8c799395e005c614be4","bytes":454236}]}
