+++ scripts/lib/edge-cfg.d.ts
@@ -1,0 +1,1 @@
+export {};
