+++ src/scripts/compile-unit-tests.ts
@@ -27,6 +27,16 @@ const {
   validateOriginAuth,
 } = require('./lib/compile-core');
 const {
+  findAdminCacheRoute,
+  collectAuthProtectedPrefixes,
+  buildAuthGateBase,
+  buildResponseCfgBase,
+  buildJwksCacheCfg,
+  buildJwtGateConfig,
+  buildSignedUrlGateConfig,
+  buildRequestCfgBase,
+} = require('./lib/edge-cfg');
+const {
   DEFAULT_ADMIN_PATH_PREFIXES,
   DEFAULT_ALLOW_METHODS,
   DEFAULT_CLEAR_SITE_DATA_TYPES,
@@ -78,6 +88,114 @@ function withEnv(key: string, value: string | undefined, fn: () => void) {
     }
   }
 }
+
+test('edge-cfg findAdminCacheRoute picks first auth or cache_control route', () => {
+  const routes = [
+    { name: 'public', match: { path_prefixes: ['/public'] } },
+    {
+      name: 'admin',
+      match: { path_prefixes: ['/admin', '/docs'] },
+      auth_gate: { type: 'static_token' },
+      response: { cache_control: 'private, max-age=60' },
+    },
+  ];
+  assert.deepStrictEqual(findAdminCacheRoute(routes), {
+    adminPathPrefixes: ['/admin', '/docs'],
+    adminCacheControl: 'private, max-age=60',
+  });
+  assert.deepStrictEqual(findAdminCacheRoute([]), {
+    adminPathPrefixes: DEFAULT_ADMIN_PATH_PREFIXES.slice(),
+    adminCacheControl: 'no-store',
+  });
+});
+
+test('edge-cfg collectAuthProtectedPrefixes unions gate prefixes', () => {
+  const prefixes = collectAuthProtectedPrefixes([
+    { protectedPrefixes: ['/admin', '/docs'] },
+    { protectedPrefixes: ['/api'] },
+    { protectedPrefixes: ['/admin'] },
+  ]);
+  assert.deepStrictEqual(prefixes, ['/admin', '/docs', '/api']);
+});
+
+test('edge-cfg buildAuthGateBase defaults admin prefixes when route has none', () => {
+  assert.deepStrictEqual(buildAuthGateBase({
+    name: 'admin',
+    match: {},
+    auth_gate: { type: 'static_token' },
+  }), {
+    name: 'admin',
+    protectedPrefixes: DEFAULT_ADMIN_PATH_PREFIXES.slice(),
+    type: 'static_token',
+  });
+});
+
+test('edge-cfg buildJwtGateConfig intersects allowed_algorithms with gate algorithm', () => {
+  const base = { name: 'api', protectedPrefixes: ['/api'], type: 'jwt' };
+  const cfg = buildJwtGateConfig({
+    algorithm: 'RS256',
+    allowed_algorithms: ['RS256', 'HS256'],
+    clock_skew_sec: 9999,
+    jwks_url: 'https://idp.example.com/jwks.json',
+  }, base);
+  assert.deepStrictEqual(cfg.allowed_algorithms, ['RS256']);
+  assert.strictEqual(cfg.clock_skew_sec, JWT_CLOCK_SKEW.max);
+  assert.strictEqual(cfg.jwks_url, 'https://idp.example.com/jwks.json');
+});
+
+test('edge-cfg buildSignedUrlGateConfig normalizes nonce_param whitespace', () => {
+  const base = { name: 'dl', protectedPrefixes: ['/download'], type: 'signed_url' };
+  const cfg = buildSignedUrlGateConfig({
+    nonce_param: '  n  ',
+    exact_path: true,
+  }, base);
+  assert.strictEqual(cfg.nonce_param, 'n');
+  assert.strictEqual(cfg.exact_path, true);
+});
+
+test('edge-cfg buildJwksCacheCfg clamps stale and negative cache seconds', () => {
+  const cfg = buildJwksCacheCfg({
+    firewall: {
+      jwks: {
+        stale_if_error_sec: 999999,
+        negative_cache_sec: 99999,
+      },
+    },
+  });
+  assert.strictEqual(cfg.staleIfErrorSec, JWKS_DEFAULTS.staleMax);
+  assert.strictEqual(cfg.negativeCacheSec, JWKS_DEFAULTS.negativeMax);
+});
+
+test('edge-cfg buildRequestCfgBase and buildResponseCfgBase share policy defaults', () => {
+  const policy = {
+    defaults: { mode: 'report' },
+    request: {
+      allow_methods: ['GET'],
+      limits: { max_query_length: 512 },
+      block: { ua_contains: ['bot'] },
+      normalize: { drop_query_keys: ['foo'] },
+    },
+    response_headers: { force_vary_auth: false },
+    routes: [{
+      name: 'admin',
+      match: { path_prefixes: ['/admin'] },
+      auth_gate: { type: 'static_token' },
+    }],
+  };
+  const requestBase = buildRequestCfgBase(policy);
+  assert.strictEqual(requestBase.mode, 'report');
+  assert.deepStrictEqual(requestBase.allowMethods, ['GET']);
+  assert.strictEqual(requestBase.maxQueryLength, 512);
+  assert.deepStrictEqual(requestBase.dropQueryKeysArray, ['foo']);
+  assert.deepStrictEqual(requestBase.uaDenyContains, ['bot']);
+
+  const authGates = [{ protectedPrefixes: ['/admin'], type: 'static_token', name: 'admin' }];
+  const responseBase = buildResponseCfgBase(policy, authGates);
+  assert.deepStrictEqual(responseBase.adminPathPrefixes, ['/admin']);
+  assert.deepStrictEqual(responseBase.authProtectedPrefixes, ['/admin']);
+  assert.strictEqual(responseBase.forceVaryAuth, false);
+  assert.strictEqual(responseBase.headers['x-content-type-options'], DEFAULT_SECURITY_HEADERS['x-content-type-options']);
+});
 
 test('policy-defaults exposes immutable runtime defaults shared by compiler targets', () => {
   assert.deepStrictEqual(DEFAULT_UA_DENY_CONTAINS, ['sqlmap', 'nikto', 'acunetix', 'masscan', 'python-requests']);
